chore(resolvers): move resolvers to separate file
