+++ lib/mongoose.js
@@ -2,7 +2,7 @@
 
 import mongoose from 'mongoose';
 
-const connectDb = handler => async (req, res) => {
+const connectDB = handler => async (req, res) => {
   if (mongoose.connections[0].readyState !== 1) {
     await mongoose.connect(process.env.DATABASE_URL, {
       useCreateIndex: true,
@@ -20,4 +20,4 @@ db.once('open', () => {
   console.log('Connected to the MongoDB.');
 });
 
-export default connectDb;
+export default connectDB;
