+++ pages/api/graphql.js
@@ -1,33 +1,7 @@
 import { ApolloServer } from 'apollo-server-micro';
-import connectDb from '../../lib/mongoose';
+import connectDB from '../../lib/mongoose';
 import typeDefs from '../../src/api/shoes/Shoes.graphql';
-import Shoes from '../../src/api/shoes/shoes';
-
-const resolvers = {
-  Query: {
-    shoes() {
-      try {
-        return [];
-      } catch (e) {
-        console.log('shoes query error:', e);
-      }
-    },
-  },
-  Mutation: {
-    async addShoe(_, { shoe }) {
-      try {
-        const newShoe = await Shoes.create(shoe);
-        return {
-          success: true,
-          message: 'tuvieja',
-          shoe: newShoe,
-        };
-      } catch (e) {
-        console.log('addShoe mutation error:', e);
-      }
-    },
-  },
-};
+import resolvers from '../../src/api/shoes/resolvers';
 
 const apolloServer = new ApolloServer({ typeDefs, resolvers });
 
@@ -38,4 +12,4 @@ export const config = {
 };
 
 const server = apolloServer.createHandler({ path: '/api/graphql' });
-export default connectDb(server);
+export default connectDB(server);
